datapipeline/pyproject.toml
[tool.poetry]
name = "datapipeline"
version = "0.1.0"
description = ""
authors = ["Bash- <Bash-@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
delta = "^0.4.2"
polars = "^0.20.17"
deltalake = "^0.16.3"
ipykernel = "^6.29.4"
google-cloud-storage = "^2.16.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.38"
langchain = "^0.1.19"
pypdf = "^4.2.0"
weaviate-client = "^4.6.0"
openai = "^1.28.1"
langchain-openai = "^0.1.6"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
